requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
autopep8==2.0.4
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
colorama==0.4.6
Django==4.2.6
executing==2.0.1
flake8==6.1.0
Flask==3.0.2
h11==0.14.0
icecream==2.1.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
lark-parser==0.12.0
MarkupSafe==2.1.5
mccabe==0.7.0
mypy==1.7.1
mypy-extensions==1.0.0
numpy==1.26.1
outcome==1.3.0.post0
packaging==24.0
pandas==2.0.3
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.16.1
PyMySQL==1.1.0
PySocks==1.7.1
pysrt==1.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
roman==4.1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
trio==0.24.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.14
types-pytz==2023.3.1.1
types-requests==2.31.0.20240125
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
vswhere==1.4.0
webdriver-manager==4.0.1
Werkzeug==3.0.1
wsproto==1.2.0
